ui 정리
